move matts utils to utils
and add another main file to distinguish between future 2d and 3d solvers and stuff
