+++ main.go
@@ -1,0 +1,6 @@
+package main
+
+func main() {
+	// In the future, we'll add command line args or other switces to run other things
+	TwoDeeAlgoSpeedStats()
+}
