+++ utils/matt/board.go
@@ -1,4 +1,4 @@
-package utils
+package matt
 
 import (
 	"strconv"
