+++ utils/matt/solve_test.go
@@ -1,4 +1,4 @@
-package utils
+package matt
 
 import (
 	"log"
