+++ utils/matt/values_test.go
@@ -1,4 +1,4 @@
-package utils
+package matt
 
 import (
 	"testing"
